Refactored reference file datamodel classes into a subdirectory.
