+++ romancal/datamodels/__init__.py
@@ -1,6 +1,6 @@
 from .core import RomanDataModel
-from .referencefile import ReferenceFileModel
-from .flat import FlatModel
+from .reference_files.referencefile import ReferenceFileModel
+from .reference_files.flat import FlatModel
 from .open_impl import open
