+++ romancal/datamodels/reference_files/referencefile.py
@@ -1,4 +1,4 @@
-from .core import RomanDataModel
+from ..core import RomanDataModel
 
 class ReferenceFileModel(RomanDataModel):
     """
